QomModulePkg: Add support for boot to fastboot using power key
Add support for boot to fastboot based on press and release time of
power key. This is to support boot to fastboot functionality for devices
which does not have volumn up/down keys. Get press and relase time
info by reading SDAM register using EFI_QCOM_PMIC_SDAM_PROTOCOL protocol
and based on range for which power key is pressed, boot to fastboot is
set.

Change-Id: Iaf186d2784c6e5dcdac52359ff028ccf8dea8649